Flyttet metrikk-logikk ut av søknads-controllere og inn i egne services
