--- src/main/kotlin/no/nav/familie/baks/mottak/søknad/AbstractSøknadMetricsService.kt
@@ -1,3 +1,0 @@
-package no.nav.familie.baks.mottak.søknad
-
-abstract class AbstractSøknadMetricsService
